backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pymongo==4.9.1
python-dotenv==1.0.1
mistralai==0.4.2
python-dateutil==2.9.0
motor==3.7.1